+++ clikt/src/commonMain/kotlin/com/github/ajalt/clikt/core/CliktCommand.kt
@@ -51,7 +51,7 @@ abstract class CliktCommand(
         private val autoCompleteEnvvar: String? = "",
         internal val allowMultipleSubcommands: Boolean = false
 ) : ParameterHolder {
-    val commandName = name ?: inferCommandName(classSimpleName())
+    val commandName = name ?: inferCommandName()
     val commandHelp = help
     val commandHelpEpilog = epilog
     internal var _subcommands: List<CliktCommand> = emptyList()
@@ -372,7 +372,9 @@ fun <T : CliktCommand> T.context(block: Context.Builder.() -> Unit): T = apply {
 
 private fun Any.classSimpleName(): String = this::class.simpleName.orEmpty().split("$").last()
 
-internal fun inferCommandName(name: String) =
-    name.removeSuffix("Command").replace(Regex("([a-z])([A-Z])")) {
+private fun CliktCommand.inferCommandName(): String {
+    val name = classSimpleName()
+    return name.removeSuffix("Command").replace(Regex("([a-z])([A-Z])")) {
         "${it.groupValues[1]}-${it.groupValues[2]}"
     }.toLowerCase()
+}